backend/work_logger_mainapp/requirement.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.9
djangorestframework==3.6.2
jira==1.0.10
oauthlib==2.0.2
ordereddict==1.1
packaging==16.8
pbr==3.0.0
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
slackclient==1.0.5
websocket-client==0.40.0
wheel==0.24.0
